# tools/audit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraderCopilot Backend",
    "version": "2.0.0"
  },
  "paths": {
...
    "/strategies/marketplace/{persona_id}/history": {
      "get": {
        "tags": [
          "Strategies",
          "strategies"
        ],
        "summary": "Get Persona History",
        "description": "Obtiene el historial de se\u00f1ales generadas por una Persona espec\u00edfica.\nFiltra por source=\"Marketplace:{persona_id}\"",
        "operationId": "get_persona_history_strategies_marketplace__persona_id__history_get",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/strategies/{strategy_id}": {
      "get": {
        "tags": [
          "Strategies",
          "strategies"
        ],
        "summary": "Get Strategy",
        "description": "Obtiene detalles de una estrategia espec\u00edfica.",
        "operationId": "get_strategy_strategies__strategy_id__get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Strategies",
          "strategies"
        ],
        "summary": "Update Strategy Config",
        "description": "Actualiza configuraci\u00f3n de una estrategia.\n\nPermite:\n- Activar/desactivar (enabled)\n- Cambiar intervalo de ejecuci\u00f3n\n- Modificar tokens/timeframes\n- Actualizar config espec\u00edfico",
        "operationId": "update_strategy_config_strategies__strategy_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/strategies/{strategy_id}/execute": {
      "post": {
        "tags": [
          "Strategies",
          "strategies"
        ],
        "summary": "Execute Strategy Manual",
        "description": "Ejecuta una estrategia manualmente (\u00fatil para testing).\n\nNo afecta al scheduler autom\u00e1tico.",
        "operationId": "execute_strategy_manual_strategies__strategy_id__execute_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteStrategyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/strategies/marketplace/create": {
      "post": {
        "tags": [
          "Strategies",
          "strategies"
        ],
        "summary": "Create Persona",
        "description": "Crea una nueva 'persona' o configuraci\u00f3n de estrategia.\nRequiere PRO.",
        "operationId": "create_persona_strategies_marketplace_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/strategies/marketplace/{persona_id}": {
      "delete": {
        "tags": [
          "Strategies",
          "strategies"
        ],
        "summary": "Delete Persona",
        "description": "Elimina una Persona personalizada del Marketplace (JSON).\nSource of Truth: JSON File directly.",
        "operationId": "delete_persona_strategies_marketplace__persona_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecuteStrategyRequest": {
        "properties": {
          "tokens": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tokens"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "default": "30m"
          },
          "context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "tokens"
        ],
        "title": "ExecuteStrategyRequest",
        "description": "Request para ejecutar estrategia manualmente."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrategyConfigUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "interval_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Seconds"
          },
          "tokens": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens"
          },
          "timeframes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeframes"
          },
          "config": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          }
        },
        "type": "object",
        "title": "StrategyConfigUpdate",
        "description": "Actualizaci\u00f3n de configuraci\u00f3n de estrategia."
      },
      "StrategyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "expected_roi": {
            "type": "string",
            "title": "Expected Roi"
          },
          "win_rate": {
            "type": "string",
            "title": "Win Rate"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          }
        },
        "type": "object",
        "required": [
          "name",
          "symbol",
          "timeframe",
          "strategy_id",
          "description",
          "risk_level",
          "expected_roi",
          "win_rate",
          "frequency"
        ],
        "title": "StrategyCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}